feat: modularized components type + fixed css scope
