+++ src/Components/WheelButton/types.ts
@@ -1,0 +1,6 @@
+type WheelButtonProps = {
+  event: () => void;
+  isDisabled: boolean;
+};
+
+export default WheelButtonProps;
