+++ src/Components/WheelSlice/types.ts
@@ -1,0 +1,7 @@
+type WheelSliceProps = {
+  championName: string;
+  sliceHeight: number;
+  color: string;
+};
+
+export default WheelSliceProps;
